+++ ToNStatTool.csproj
@@ -118,12 +118,30 @@
     <Compile Include="ToNStatTool.Designer.cs">
       <DependentUpon>ToNStatTool.cs</DependentUpon>
     </Compile>
+    <Compile Include="Forms\MainFormControls.cs">
+      <SubType>Form</SubType>
+    </Compile>
+    <Compile Include="Forms\MainFormEventHandlers.cs">
+      <SubType>Form</SubType>
+    </Compile>
+    <Compile Include="Forms\MainFormUIUpdaters.cs">
+      <SubType>Form</SubType>
+    </Compile>
+    <Compile Include="Helpers\ControlHelper.cs" />
+    <Compile Include="Services\SoundPlayer.cs" />
+    <Compile Include="Services\WarningUserManager.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="WebSocketClient.cs" />
     <Compile Include="Controls\DoubleBufferedListView.cs">
       <SubType>Component</SubType>
     </Compile>
     <Content Include="icon.ico" />
+    <Content Include="item.mp3">
+      <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
+    </Content>
+    <Content Include="masterchange.mp3">
+      <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
+    </Content>
     <Content Include="player_join.mp3">
       <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
     </Content>
